Accept only 1 client per round, make sock be stdin
These two are linked, because systemd only applies its ‘NonBlocking’
directive to file descriptors ≥ #3. (That’s lame if you ask me.)

Conceptually, the listening socket _is_ the input door, so it makes sense
to have it be stdin (filedes #0.)
And that way, the “close stdin to make it filedes be reusable” logic
in main() can go away*.

It’s anyway better not to accept too many new clients per round.
The listening socket backlog (defined in sshtarp.socket, btw reduced
in this commit) previously helped in case of a burst, but not totally.
Just one new client every $SLEEP_TIME seconds is sane enough, and this
provides better safety in case of DoS attempt.

--

*: This logic is generally dangerous: closing/messing with std{in,out,err}
makes some programs, and even some libraries (!), go nuts.
But it was OK here, and I do not like to waste resources, be them file
descriptors or anything else.