Add more testing clients, using popular SSH libs
“Testing” here means checking that a client using such and such SSH
implementation does hang when trying to communicate with sshtarp.
It _should_ hang, for hours: sshtarp is a tarpit!

To set up testing, first run `sudo make service` (on a Linux+systemd
machine), to compile+start sshtarp.
Then run `./test_$CLIENTLIB` (or, in the case of the Golang client,
`go run test_goxssh.go`).

A way to get the dependencies for each client is given in a comment
in their file. Assuming an Ubuntu machine (probably works on Debian too).