Advertise more algorithms, to fool more clients
So, some of the testing clients added in the previous commit (e34c8dfb9)
were not satisfied with the algorithms previously advertised.

The golang SSH lib & Net::SSH::Perl, in their default config at least,
are not content with “aes128-cbc” as the only encryption cipher. Adding
“aes256-ctr” suits them.

Net::SSH::Perl by default is not OK with “hmac-sha1” as the MAC. So add
“hmac-sha2-256”. Although it’s unlikely that some bots use Net::SSH::Perl
(it’s a pure Perl implementation of SSH, not a binding).

For good measure only, add “ssh-dss” as an advertised host key algo.
This leads to crossing an 8-byte threshold, so add “zlib” as a
compression method. Otherwise the space is taken by more padding.
(It’s better to fill the space with some algo, because the client has to
record the full KEYXINIT message in order to then hash it; see
https://tools.ietf.org/html/rfc4253#section-8 — and if I’m not mistaken,
the padding itself is part of the _packet_, not part of the _message_,
hence it does not have to be recorded+hashed by the client.)