Don’t set first_kex_packet_follows in KEXINIT msg
It was previously set (to 1 = true), ± thinking that this field was
concerned with the transport layer (TCP), because indeed in the same
TCP/IP packet we are bundling a banner message, a SSH_MSG_KEYXINIT,
and the start of a (bogus) SSH_MSG_KEXDH_REPLY.
But that reasoning was wrong; this field is concerned with the
application layer (SSH), so to speak. (It’s for when you already know
the other peer, hence also the algos you can/will both agree to use.)

It was no big deal to set it, but:
 “After receiving the SSH_MSG_KEXINIT packet from the other side, each
 party will know whether their guess was right. If the other party's
 guess was wrong, and this field was TRUE, the next packet MUST be
 silently ignored” — https://tools.ietf.org/html/rfc4253#section-7.1

And indeed the next packet (the bogus SSH_MSG_KEXDH_REPLY) gets ignored,
at least by OpenSSH:
 […]
 debug2: first_kex_follows 1
 […]
 debug1: expecting SSH2_MSG_KEXDH_REPLY
 debug3: receive packet: type 32
 debug2: skipped packet (type 32)
 Bad packet length 1208984698.
 debug3: send packet: type 1
 ssh_dispatch_run_fatal: Connection to 127.0.0.1 port 22: message authentication code incorrect

When first_kex_packet_follows is not set, here’s what we get instead:
 […]
 debug2: first_kex_follows 0
 […]
 debug1: expecting SSH2_MSG_KEXDH_REPLY
 debug3: receive packet: type 32
 kex protocol error: type 32 seq 1
 debug3: send packet: type 3
 Bad packet length 4287849664.
 debug3: send packet: type 1
 ssh_dispatch_run_fatal: Connection to 127.0.0.1 port 22: message authentication code incorrect

That’s better, because we want the client to waste some energy trying
to parse the garbage we sent it. And who knows, this may make it crash,
or fool it for even longer, etc.
Also, notice how OpenSSH sends an extra message in this case (type 3,
SSH_MSG_UNIMPLEMENTED). Having clients waste more bandwidth is good too!