Increase to 32 bytes the size of BOGUS_DATA
It’s criminal to syscall write() for tiny amounts of data, and it’s
criminal to send TCP/IP packets where the TCP+IP headers are bigger than
the actual payload. (By “criminal” I mean “terribly inefficient/wasteful”).

But well, we do it on purpose here, that’s the point of a tarpit. Trade
less time where the client is stuck for a bit more efficient use of
resources, though.

Make the extra bytes as fishy as the KEXINIT cookie & padding; make it
look like the machine is compromised, and someone is subtly exfiltrating
/etc/passwd entries.