Don’t poll for clients writability
There is no need: they are either ready to receive more junk data,
or so slow that they deserve to be close()d. We only send 32 bytes
of data every $SLEEP_TIME seconds.

Just record the clients file descriptors, loop over them every round
(every $SLEEP_TIME seconds), and call write(). The write either succeeds
or we close the client, very probably because it closed the connection
on its end. (We don’t have to explicitely care about why/what failed;
we can just log the value of errno and move on.)

A file descriptor is an int, so (on an x64 machine) 4 bytes. That’s a
third of the size of a ‘struct epoll_event’ (12 bytes). Less memory use
is always cool, that also means less pressure on the CPU caches.
And there are now less bytes crossing the kernel/user-space border every
round.

We still have to poll to see when/if they are new clients to accept().
We can simply resort to poll() for that; it’s simpler than epoll(), and
standardized by POSIX (while epoll is Linux specific). That will help
portability to the BSDs (& co.), which is a mid-term goal.

Regarding the clients file descriptors, we store them in a simple array,
statically allocated to its max capacity ($MAX_CLIENTS slots). We did
that before too, with the ‘struct epoll_event’ array. However, we now
have more management to do.
Given a client fd, we compute its index in the array using substraction.
That’s easy but sloppy; an assertion in welcome_new_client prevents out
of bounds access. Worse is better.
The array can be sparse, we manage that using the “tombstone trick”, seen
in some hash tables.
See http://www.craftinginterpreters.com/hash-tables.html#deleting-entries
We also keep track of where the last occupied slot here, using another
special value: “END”, equals to 0. That makes the array initialization
easy, but that’s a bit sloppy too, given that 0 is a valid fd number
(it’s stdin).

Gone is the $COMA_TIME logic: now if $MAX_CLIENTS is reached, we instead
free (close) some of the clients. That’s arguably better.