Ensure expected filedes⇒ indexes in clients array
So the previous commit (e36f25b035) message mentions that:
> Given a client fd, we compute its index in the array using substraction.
> That’s easy but sloppy […]

And indeed it is.

Imagine someone is launching sshtarp with stdout closed, e.g. like that:
 $ ./sshtarp >&-

Then the first client we accept() will get fd #1 (closed fds get reused).

Using (fd − FIRST_CLIENT_FD) as the formula to get from an fd to an index
in the clients array, that leads to an index of (1 − 3) = −2, which is,
obviously, out of bounds.

The previous commit message also says that:
> an assertion in welcome_new_client prevents out of bounds access.

And indeed, the program would die() in that case. But it could have died
earlier, and with a better explanation.

That’s now the case with this commit. (The “better explanation” being
“stdout: Bad file descriptor”, which makes sense if you speak Unix.)

N.B.: if the user doesn’t care about log messages, she/he can still
_redirect_ stdout to /dev/null (but not _close_ it). The present state
of sshtarp illustrates why null-redirecting is generally the right thing
to do. (See also the footnote in commit bd5e791885.)

--

Now imagine someone is launching sshtarp like that:
 $ ./sshtarp 4<somefile

That is, with fd #4 opened for reading ‘somefile’. Crazy and pointless
way to start the program, but possible way. (And sometimes “crazy” ways
actually are übersmart — people know better than you why/what they do.)

Here the first client gets assigned fd #3, and the second one fd #5.
We get non-continuous indexes, and a hole in the clients array.
Because of the way the array is initialized, that hole is actually taken
by the special value “END”. The writing loop in handle_clients stops as
soon as it sees this marker. So we never write to (or close!) anyone but
the first client!

That’s fixed with this commit: file descriptors intended for client
sockets get closed before the main infinite loop starts.

--

About all this, the previous commit message concludes:
> Worse is better.

A reference to http://dreamsongs.com/RiseOfWorseIsBetter.html,
particularly: “The worse-is-better philosophy [deems that] It is slightly
better to be simple than correct.”

The opposite “the-right-thing” approach would have been to completely
dissociate fds and indexes. Probably use a linked list to represent the
list of clients. That would have mean more code and more cache trashing,
to be right in 100% instead of 99% of the cases.

And with this commit I think we are right in 99.5% of the cases.
To the best of my knowledge, it’s not possible to externally mess with
the file descriptors of a process (i.e. a process A cannot open/close
the fds of a process B). The only issue that remains (that I can think
of) is if the code is modified & the programmer forgets/ignores about
all this.

--

The “trick” we use to check if a fd is opened or closed is to poll() it.
If you get back POLLNVAL, then it’s closed. (It seems sane to assume
there’s no other reason why you’d get POLLNVAL.)

There are other ways to check. fcntl() is one. But here we have to check
in bulk, so one call to poll() is better. Bonus point: no need to allow
yet another syscall in sshtarp.service

Related: the closefrom() syscall, available in the BSDs & Solaris but not
on Linux. See https://man.openbsd.org/closefrom.2 and
https://twitter.com/bcantrill/status/661321262258479104