Desynchronize write() and accept()
Previously in every round of the “main infinite loop”, we were doing:
  — a write($BOGUS_DATA) to every established clients
  — an accept() of any queued (new) client
And the interval between rounds was $SLEEP_TIME seconds.

(“Clients”, here & in general in this project, mean “malevolent bots”.)

But that implied conflicting requirements on what $SLEEP_TIME should be:
  — the interval between writes should be as high as possible, to keep
  the clients stuck for as long as possible, with the least amount of
  work for us. We just don’t want to trigger a timeout on their side.
  — on the other hand, we want to troll as many clients as sensible, so
  we want the interval between accepts() to be quite low. Moreover, I
  suspect that many clients don’t wait much to see if they can connect,
  even if they tolerate, after that, much longer delays.

So instead of settling on a “middle-ground” $SLEEP_TIME that does not
really satisfy anything, this commit replaces it by two independent
values, $WRITES_INTERVAL and $ACCEPTS_INTERVAL. And in a loop round, we
now do either a “write pass”, or an “accept pass”, or, sometimes, both.

--

I was dreadful this change was going to make the codebase transition
from “obviously error-free” to “no obvious error”*. But it turns out
that after some thinking and trials, a simple, clear solution could be
devised. One key is to deal as much as possible in absolute times.

By the way:
 “Resist the temptation to implement a sleep for a fixed amount of time
  by using the return value of sleep, when nonzero, to call sleep again.
  This will work with a certain amount of accuracy as long as signals
  arrive infrequently. […]
  Instead, compute the calendar time at which the program should stop
  waiting, and keep trying to wait until that calendar time. This won’t
  be off by more than a second.”
 — https://www.gnu.org/software/libc/manual/html_node/Sleeping.html

(We are, thankfully, not concerned by signals; we don’t handle them,
and let the program exits if it receives any. Signals stink.)

--

There was a mistake in time keeping previously, introduced in e36f25b03:
> case 1:
>   welcome_new_client(LISTEN_FD, clients);
>   sleep(time(NULL) - start);

should have ± been:
> […]
>   sleep(SLEEP_TIME - (time(NULL) - start));

The time spent blocked in poll() should indeed be measured … for it
to be substracted from $SLEEP_TIME. I forgot that substraction 🤦🙊

--

*: “There are two ways of constructing a software design: One way is to
make it so simple that there are obviously no deficiencies, and the other
way is to make it so complicated that there are no obvious deficiencies.
The first method is far more difficult.” — C.A.R. Hoare in “The Emperor’s
Old Clothes” http://zoo.cs.yale.edu/classes/cs422/2011/bib/hoare81emperor.pdf