Guard our dprintf() from overflowing its buffer
Better (truly) safe than sorry. Previously we were only “safe” because
the buffer was purposely oversized.

If an %u or %s would make the buffer overflow, it is silently discarded.
Can this be potentially surprising, even misleading? Yes. But so can a
truncated number/string. And I don’t want to pointlessly complicate all
this; overflows will “never happen”™.