Increase SendBuffer in trassh.socket (big RSA key)
A handshake with a big (10+ kibibits) RSA key will be several kilobytes
long. In that case, it makes sense to have the socket send buffer be
bigger; big enough to avoid the kernel having to flush-then-refill it.
Otherwise, it’s not much of a waste; the kernel may be allocating socket
buffers in whole pages (generally 4 KiB) anyway (?)