Install trassh.dat in a righter place
It turns out that a fresh FreeBSD install has no `/var/local` directory.

Anyway, `/usr/local/share` (the new location) is a better place for
`trassh.dat` to be, since it is a static, read-only data file.

> /var contains variable data files. This includes spool directories
> and files, administrative and logging data, and transient and
> temporary files. […]
> Everything that once went into /usr that is written to during system
> operation (as opposed to installation and software maintenance) must
> be in /var.
— https://refspecs.LinuxFoundation.org/FHS_3.0/fhs/ch05.html#purpose31

> Any program or package which contains or requires data that doesn't
> need to be modified should store that data in /usr/share (or
> /usr/local/share, if installed locally). It is recommended that a
> subdirectory be used in /usr/share for this purpose. Applications
> using a single file may use /usr/share/misc.
— https://refspecs.LinuxFoundation.org/FHS_3.0/fhs/ch04s11.html#purpose26

`/usr/local/share/misc` may have been an even better place for
`trassh.dat`, but no such directory exists on Ubuntu or FreeBSD 😐
Not creating a subdirectory in `/usr/local/share` because overkill.