Don’t read past received data in trassh.c 😱
Not segfault-inducing here, because the receive buffer is bigger than
BANNER_MAXLEN. Still a serious defect: a newline character that was
_not_ actually sent by the client could potentially be found in the
BANNER_MAXLEN first bytes of the buffer.
The buffer is an “automatic storage” (i.e.: stack-allocated) variable,
so by default filled with random junk, per the C standard (see below.)
In that junk, there could be a newline!

> If an object that has automatic storage duration is not initialized
> explicitly, its value is indeterminate.
— http://www.open-std.org/jtc1/sc22/wg14/www/docs/n1256.pdf#page=138

(Regarding codestyle: yes, that `if` got quite hairy, but it will do.)