Hard require TCP/IP sockets in trassh.c
The SSH protocol is layered on top of TCP/IP. Even for testing purposes,
it’s not hard to set up TCP connections (see the `client_*` programs.)

Sacrificing the flexibility to run with virtually any kind of sockets
makes for simpler code. A pesky `goto` & associated label got removed 😌
And this flexibility was not exercised/tested anyway, so potentially not
properly working!